Re-use make_cfg
