+++ examples/settings.py
@@ -2,6 +2,10 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+import habitat_sim
+import habitat_sim.agent
+import habitat_sim.bindings as hsim
+
 
 default_sim_settings = {
     # settings shared by example.py and benchmark.py
@@ -25,3 +29,63 @@ default_sim_settings = {
     "goal_position": [-9.423, 0.072_447, -0.373],
     "goal_headings": [[0.0, -0.471_395, 0.0, 0.881_922], [0.0, 1.0, 0.0, 0.0]],
 }
+
+# build SimulatorConfiguration
+def make_cfg(settings):
+    sim_cfg = hsim.SimulatorConfiguration()
+    sim_cfg.gpu_device_id = 0
+    sim_cfg.scene.id = settings["scene"]
+
+    # define default sensor parameters (see src/esp/Sensor/Sensor.h)
+    sensors = {
+        "color_sensor": {  # active if sim_settings["color_sensor"]
+            "sensor_type": hsim.SensorType.COLOR,
+            "resolution": [settings["height"], settings["width"]],
+            "position": [0.0, settings["sensor_height"], 0.0],
+        },
+        "depth_sensor": {  # active if sim_settings["depth_sensor"]
+            "sensor_type": hsim.SensorType.DEPTH,
+            "resolution": [settings["height"], settings["width"]],
+            "position": [0.0, settings["sensor_height"], 0.0],
+        },
+        "semantic_sensor": {  # active if sim_settings["semantic_sensor"]
+            "sensor_type": hsim.SensorType.SEMANTIC,
+            "resolution": [settings["height"], settings["width"]],
+            "position": [0.0, settings["sensor_height"], 0.0],
+        },
+    }
+
+    # create sensor specifications
+    sensor_specs = []
+    for sensor_uuid, sensor_params in sensors.items():
+        if settings[sensor_uuid]:
+            sensor_spec = hsim.SensorSpec()
+            sensor_spec.uuid = sensor_uuid
+            sensor_spec.sensor_type = sensor_params["sensor_type"]
+            sensor_spec.resolution = sensor_params["resolution"]
+            sensor_spec.position = sensor_params["position"]
+            if not settings["silent"]:
+                print("==== Initialized Sensor Spec: =====")
+                print("Sensor uuid: ", sensor_spec.uuid)
+                print("Sensor type: ", sensor_spec.sensor_type)
+                print("Sensor position: ", sensor_spec.position)
+                print("===================================")
+
+            sensor_specs.append(sensor_spec)
+
+    # create agent specifications
+    agent_cfg = habitat_sim.agent.AgentConfig()
+    agent_cfg.sensor_specifications = sensor_specs
+    agent_cfg.action_space = {
+        "move_forward": habitat_sim.agent.ActionSpec(
+            "move_forward", habitat_sim.agent.ActuationSpec(amount=0.25)
+        ),
+        "turn_left": habitat_sim.agent.ActionSpec(
+            "turn_left", habitat_sim.agent.ActuationSpec(amount=10.0)
+        ),
+        "turn_right": habitat_sim.agent.ActionSpec(
+            "turn_right", habitat_sim.agent.ActuationSpec(amount=10.0)
+        ),
+    }
+
+    return habitat_sim.Configuration(sim_cfg, [agent_cfg])
